changed message loading issue
